+++ languagetool-core/src/main/java/org/languagetool/JLanguageTool.java
@@ -31,7 +31,6 @@ import org.languagetool.markup.AnnotatedText;
 import org.languagetool.markup.AnnotatedTextBuilder;
 import org.languagetool.markup.TextPart;
 import org.languagetool.rules.*;
-import org.languagetool.rules.neuralnetwork.Word2VecModel;
 import org.languagetool.rules.patterns.*;
 import org.languagetool.rules.spelling.SpellingCheckRule;
 import org.languagetool.tools.LoggingTools;
@@ -579,19 +578,6 @@ public class JLanguageTool {
   }
 
   /**
-   * Activate rules that depend on pre-trained neural network models.
-   *
-   * @param modelDir root dir of exported models
-   * @since 4.4
-   */
-  public void activateNeuralNetworkRules(File modelDir) throws IOException {
-    ResourceBundle messages = getMessageBundle(language);
-    List<Rule> rules = language.getRelevantNeuralNetworkModels(messages, modelDir);
-    userRules.addAll(rules);
-    ruleSetCache.clear();
-  }
-
-  /**
    * Activate rules that depend on a language model. The language model currently
    * consists of Lucene indexes with ngram occurrence counts.
    *
@@ -643,21 +629,6 @@ public class JLanguageTool {
     transformRules(enhanced, builtinRules);
     transformRules(enhanced, userRules);
     ruleSetCache.clear();
-  }
-
-  /**
-   * Activate rules that depend on a word2vec language model.
-   *
-   * @param indexDir directory with a subdirectories like 'en', each containing dictionary.txt and final_embeddings.txt
-   * @since 4.0
-   */
-  public void activateWord2VecModelRules(File indexDir) throws IOException {
-    Word2VecModel word2vecModel = language.getWord2VecModel(indexDir);
-    if (word2vecModel != null) {
-      ResourceBundle messages = getMessageBundle(language);
-      List<Rule> rules = language.getRelevantWord2VecModelRules(messages, word2vecModel);
-      userRules.addAll(rules);
-    }
   }
 
   /**
@@ -1306,25 +1277,6 @@ public class JLanguageTool {
       printIfVerbose(analyzedSentence.toString());
       printIfVerbose(analyzedSentence.getAnnotations());
     }
-  }
-
-  /**
-   * @deprecated use {@link #performCheck(List, List, RuleSet, ParagraphHandling, AnnotatedText, RuleMatchListener, Mode, Level, boolean)}
-   */
-  @Deprecated
-  protected CheckResults performCheck(List<AnalyzedSentence> analyzedSentences, List<String> sentences,
-                                         List<Rule> allRules, ParagraphHandling paraMode, AnnotatedText annotatedText, Mode mode, Level level) throws IOException {
-    List<Rule> nonIgnored = allRules.stream().filter(r -> !ignoreRule(r)).collect(Collectors.toList());
-    return performCheck(analyzedSentences, sentences, nonIgnored, paraMode, annotatedText, null, mode, level, true);
-  }
-
-  /**
-   * @deprecated use {@link #performCheck(List, List, RuleSet, ParagraphHandling, AnnotatedText, RuleMatchListener, Mode, Level, boolean)}
-   * @since 3.7
-   */
-  protected CheckResults performCheck(List<AnalyzedSentence> analyzedSentences, List<String> sentenceTexts,
-                                         List<Rule> allRules, ParagraphHandling paraMode, AnnotatedText annotatedText, RuleMatchListener listener, Mode mode, Level level, boolean checkRemoteRules) throws IOException {
-    return performCheck(analyzedSentences, sentenceTexts, RuleSet.plain(allRules), paraMode, annotatedText, listener, mode, level, checkRemoteRules);
   }
 
   /**