+++ languagetool-core/src/test/java/org/languagetool/rules/patterns/PatternRuleTest.java
@@ -328,7 +328,6 @@ public class PatternRuleTest extends AbstractPatternRuleTest {
     for (Rule rule : rules) {
       if (rule instanceof AbstractPatternRule) {
         AbstractPatternRule apRule = (AbstractPatternRule) rule;
-        List<PatternToken> patternTokens = apRule.getPatternTokens();
         List<Match> suggestionMatches = new ArrayList<>();
         if (apRule.getSuggestionMatches() != null) {
           suggestionMatches.addAll(apRule.getSuggestionMatches());
@@ -463,9 +462,20 @@ public class PatternRuleTest extends AbstractPatternRuleTest {
       if (msg.toLowerCase().contains("tbd")) {
         fail("Unfinished message (contains 'tbd') of rule " + rule.getFullId() + ": '" + msg + "'");
       }
-      //if (msg.matches(".*[^\"'>)?!.]$")) {
-      //  System.err.println("Warning: Message of " + rule.getFullId() + " doesn't end in [.!?]: " + msg);
-      //}
+      if (lang.getShortCode().matches("de|en|fr|es|nl")) {  // not yet 'pt' due to many matches there
+        if (!msg.trim().equals(rule.getMessage())) {
+          System.err.println("*** WARNING: Message of rule " + rule.getFullId() + " starts or ends with spaces: '" + rule.getMessage() + "'");
+        }
+      }
+      if (lang.getShortCode().equals("de") && !msg.equals("Failing for testing purposes")) {
+        if (msg.trim().endsWith("!")) {
+          fail("Message ends in '!' for rule " + rule.getFullId() + ": '" + msg + "'");
+        }
+        if (!msg.trim().matches(".*[.?)'\"]$")) {
+          fail("Message doesn't end with [.?)'\"] for rule " + rule.getFullId() + ": '" + msg + "'");
+          //System.out.println("Message doesn't end with [.?)'\"] for rule " + rule.getFullId() + ": '" + msg + "'");
+        }
+      }
     }
   }